requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.1
MarkupSafe==2.1.5
python-docx==1.1.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
